refactor: split server.go command dispatch into handlers.go
